requirements.txt
numpy
setuptools
pytz
nose
opencv-python
click
requests
pypdf
Pillow
pdf2image
pytesseract